requirements.txt
aiosqlite==0.19.0
anyio==3.7.1
asyncpg==0.31.0
attrs==25.4.0
bcrypt==4.1.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==1.10.13
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-jose==3.5.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.31.0
rsa==4.9.1
selenium==4.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.51
starlette==0.27.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.24.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.1.0